VGS命令を廃止 & INT命令を追加
