+++ .gitignore
@@ -63,7 +63,7 @@ tp_cmp2_a
 tp_cmp2_b
 tp_cmp2_c
 tp_cmp2_d
-tp_vgs
+tp_int
 vgsasm
 asmtest
 tp.bin
