+++ src/asm/vgsasm_parse.c
@@ -156,8 +156,8 @@ int parse_operation(struct line_data* line, int len)
             if (parse_shift(line, i, SHIFT_LEFT)) error_count++;
         } else if (0 == strcasecmp(line[i].token[0], "SHR")) {
             if (parse_shift(line, i, SHIFT_RIGHT)) error_count++;
-        } else if (0 == strcasecmp(line[i].token[0], "VGS")) {
-            if (parse_vgs(line, i)) error_count++;
+        } else if (0 == strcasecmp(line[i].token[0], "INT")) {
+            if (parse_int(line, i)) error_count++;
         } else if (0 == strcasecmp(line[i].token[0], "ADD")) {
             if (parse_acu(line, i, ACU_ADD)) error_count++;
         } else if (0 == strcasecmp(line[i].token[0], "SUB")) {
