--- src/cpu/vgsapi.c
@@ -1,6 +1,0 @@
-#include "vgscpu_internal.h"
-
-int vgsapi_noop(struct vgscpu_context* c)
-{
-    return 0;
-}
